+++ tools/packman/bootstrap/install_package.py
@@ -142,7 +142,7 @@ def generate_sha256_for_file(file_path: Union[str, os.PathLike]) -> str:
 
 
 def install_common_module(package_path, install_path):
-    COMMON_SHA256 = "ec6d8f2b2583a0f63dbc1e84366e198cc875afaec29f7d36e343d8887289ed43"
+    COMMON_SHA256 = "d24e3825b6fd4ee55758d271bf8ce0011538e70a281d4642893fbd38edaab035"
     package_sha256 = generate_sha256_for_file(package_path)
     if package_sha256 != COMMON_SHA256:
         raise RuntimeError(